python_basics/Iterators-And-Streams/Transform/pyproject.toml
[tool.poetry]
name = "yourname-wf-basic"
version = "0.2.0"
description = "A pipeline processing system"
authors = ["Your Name <yourname@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
pyyaml = "^5.4.1"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
